[dev] lite: drop the low-value tests and merge the rest
Two test files went: the dispatcher permutation index test, whose change
measured zero and only removed an environment-variable knob, and the DS4
cross-entropy fusion wiring test, which the review panel itself flagged as
wiring without numerical proof. The three CSA/expert fusion files merged into
one, dropping duplicated imports and fixtures.

Long comment blocks in the production files were compressed to their first
sentence. The measurements and causal chains they carried are in the pull
request, where they can be read once and will not go stale in the source.

The nine negative controls are untouched -- they are why the review panel found
no blockers, and they are the part of a test that a later refactor cannot
quietly satisfy.
